refactor: 공유 패키지 분리 및 변경 (#133)
