+++ src/main/java/site/gachontable/domain/admin/port/in/AdminRegister.java
@@ -1,7 +1,7 @@
 package site.gachontable.domain.admin.port.in;
 
 import site.gachontable.presentation.admin.dto.request.AdminRegisterRequest;
-import site.gachontable.domain.shared.dto.response.RegisterResponse;
+import site.gachontable.presentation.shared.dto.response.RegisterResponse;
 
 public interface AdminRegister {
