+++ src/main/java/site/gachontable/domain/admin/port/in/UpdateStatus.java
@@ -2,7 +2,7 @@ package site.gachontable.domain.admin.port.in;
 
 import site.gachontable.presentation.admin.dto.request.UpdateStatusRequest;
 import site.gachontable.infra.security.principal.AuthDetails;
-import site.gachontable.domain.shared.dto.response.RegisterResponse;
+import site.gachontable.presentation.shared.dto.response.RegisterResponse;
 
 public interface UpdateStatus {
