+++ src/main/java/site/gachontable/domain/member/service/UserLoginImpl.java
@@ -3,8 +3,8 @@ package site.gachontable.domain.member.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import site.gachontable.domain.shared.Role;
-import site.gachontable.domain.shared.exception.PasswordNotMatchException;
+import site.gachontable.presentation.shared.Role;
+import site.gachontable.presentation.shared.exception.PasswordNotMatchException;
 import site.gachontable.domain.member.domain.User;
 import site.gachontable.domain.member.port.out.UserRepository;
 import site.gachontable.domain.member.exception.UserNotFoundException;
