--- src/main/java/site/gachontable/gachontablebe/domain/shared/dto/request/RefreshRequest.java
@@ -1,4 +1,0 @@
-package site.gachontable.gachontablebe.domain.shared.dto.request;
-
-public record RefreshRequest(String refreshToken) {
-}
