+++ src/main/java/site/gachontable/presentation/admin/api/AdminController.java
@@ -19,8 +19,8 @@ import site.gachontable.presentation.admin.dto.response.WaitingInfosResponse;
 import site.gachontable.domain.admin.type.Status;
 import site.gachontable.infra.security.principal.AuthDetails;
 import site.gachontable.presentation.admin.dto.request.PubManageRequest;
-import site.gachontable.domain.shared.dto.request.RefreshRequest;
-import site.gachontable.domain.shared.dto.response.RegisterResponse;
+import site.gachontable.presentation.shared.dto.request.RefreshRequest;
+import site.gachontable.presentation.shared.dto.response.RegisterResponse;
 import site.gachontable.domain.waiting.service.GetWaitings;
 import site.gachontable.independent.error.ErrorResponse;
 import site.gachontable.infra.security.jwt.JwtProvider;
