+++ src/main/java/site/gachontable/presentation/admin/dto/response/SeatingsResponse.java
@@ -2,12 +2,12 @@ package site.gachontable.presentation.admin.dto.response;
 
 import lombok.Builder;
 import site.gachontable.domain.seating.domain.Seating;
-import site.gachontable.domain.shared.Table;
+import site.gachontable.presentation.shared.Table;
 
 import java.util.List;
 import java.util.UUID;
 
-import static site.gachontable.domain.shared.DateTimeFormatters.WITH_WEEKDAY;
+import static site.gachontable.presentation.shared.DateTimeFormatters.WITH_WEEKDAY;
 
 public record SeatingsResponse(List<SeatingResponse> seatings) {
