+++ src/main/java/site/gachontable/presentation/member/api/UserController.java
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import site.gachontable.domain.shared.dto.request.RefreshRequest;
-import site.gachontable.domain.shared.dto.request.TestRegisterRequest;
-import site.gachontable.domain.shared.dto.response.RegisterResponse;
+import site.gachontable.presentation.shared.dto.request.RefreshRequest;
+import site.gachontable.presentation.shared.dto.request.TestRegisterRequest;
+import site.gachontable.presentation.shared.dto.response.RegisterResponse;
 import site.gachontable.presentation.member.dto.request.UserLoginRequest;
 import site.gachontable.domain.member.port.in.UserLogin;
 import site.gachontable.domain.member.port.in.UserRegister;
