+++ src/main/java/site/gachontable/presentation/shared/dto/request/RefreshRequest.java
@@ -1,0 +1,4 @@
+package site.gachontable.presentation.shared.dto.request;
+
+public record RefreshRequest(String refreshToken) {
+}
