+++ src/main/java/site/gachontable/presentation/shared/exception/PasswordNotMatchException.java
@@ -1,0 +1,11 @@
+package site.gachontable.presentation.shared.exception;
+
+
+import site.gachontable.independent.type.ErrorCode;
+import site.gachontable.independent.error.exception.ServiceException;
+
+public class PasswordNotMatchException extends ServiceException {
+    public PasswordNotMatchException() {
+        super(ErrorCode.PASSWORD_NOT_MATCH);
+    }
+}
