+++ src/main/java/site/gachontable/presentation/waiting/dto/request/RemoteWaitingRequest.java
@@ -1,6 +1,6 @@
 package site.gachontable.presentation.waiting.dto.request;
 
-import site.gachontable.domain.shared.Table;
+import site.gachontable.presentation.shared.Table;
 
 public record RemoteWaitingRequest(Integer pubId,
                                    Table tableType) {
